+++ tpm/tss2/tss2.go
@@ -1,4 +1,4 @@
-package storage
+package tss2
 
 import (
 	"encoding/asn1"
